Change seeders location
